+++ packages/expo-dev-menu/vendored/react-native-reanimated/ios/DevMenuRNGestureHandlerStateManager.h
@@ -1,5 +1,7 @@
-@protocol DevMenuRNGestureHandlerStateManager
-
-- (void)setGestureState:(int)state forHandler:(int)handlerTag;
-
-@end
+// This interface is located in the gesture handler package.
+// We have to remove it from reanimated to avoid duplicate files error.
+//@protocol DevMenuRNGestureHandlerStateManager
+//
+//- (void)setGestureState:(int)state forHandler:(int)handlerTag;
+//
+//@end